requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-rest-framework-async==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pyaes==1.6.1
pyasn1==0.6.0
PyJWT==2.8.0
pyTelegramBotAPI==4.21.0
python-telegram-bot==21.3
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.0
Telethon==1.36.0
typing_extensions==4.12.2
urllib3==2.2.2
